+++ internal/app/app.go
@@ -18,6 +18,7 @@ func Run(args []string, stdout, stderr io.Writer) int {
 	searchCmd := commands.NewSearchCommand()
 	autocompleteCmd := commands.NewAutocompleteCommand()
 	cacheCmd := commands.NewCacheCommand()
+	snapshotCmd := commands.NewSnapshotCommand()
 
 	var ordered []commands.Command
 	lookup := map[string]commands.Command{}
@@ -33,7 +34,7 @@ func Run(args []string, stdout, stderr io.Writer) int {
 
 	helpCmd := commands.NewHelpCommand(usage, lookupFn)
 
-	ordered = []commands.Command{helpCmd, authTokenCmd, geocodeCmd, reverseCmd, directionsCmd, searchCmd, autocompleteCmd, cacheCmd, pingCmd, versionCmd}
+	ordered = []commands.Command{helpCmd, authTokenCmd, geocodeCmd, reverseCmd, directionsCmd, searchCmd, autocompleteCmd, cacheCmd, snapshotCmd, pingCmd, versionCmd}
 	for _, cmd := range ordered {
 		lookup[cmd.Name] = cmd
 	}